feat: activityTabView에서 사용되던 search기능을 activityTimer로 옮김
